# perf(int_extension): rewrite `isPrime` method with more performant iteration