--- a/lib/utils/int_extension.dart
+++ b/lib/utils/int_extension.dart
@@ -1,16 +1,27 @@
-import 'dart:math';
+final Map<int, bool> primes = {
+  1: false,
+  2: true,
+};
 
 extension IntExtension on int {
   bool get isPrime {
-    if (this <= 1) return false;
-    if (this == 2) return true;
-    if (isEven) return false;
+    final cachedPrime = primes[this];
+    if (cachedPrime != null) return cachedPrime;
 
-    final maxDivisor = sqrt(this).floor();
-    for (var i = 3; i <= maxDivisor; i += 2) {
-      if (this % i == 0) return false;
+    if (this <= 1 || isEven) return primes[this] = false;
+
+    for (var i = 3;
+        // Do not excede the square root of num.
+        i * i <= this;
+        // All prime numbers are 1 or 5 modulo 6.
+        // Since we start with 3, this will do:
+        // 3 -> 5 -> 7 -> 11 ... +2 -> +4 -> +2 -> +4 ...
+        i = i % 6 == 1 ? i + 4 : i + 2) {
+      if (this % i == 0) {
+        return primes[this] = false;
+      }
     }
 
-    return true;
+    return primes[this] = true;
   }
 }