--- a/lib/utils/int_extension.dart
+++ b/lib/utils/int_extension.dart
@@ -7,7 +7,12 @@ extension IntExtension on int {
     if (isEven) return false;
 
     final maxDivisor = sqrt(this).floor();
-    for (var i = 3; i <= maxDivisor; i = i % 6 == 1 ? i + 4 : i + 2) {
+    for (var i = 3;
+        i <= maxDivisor;
+        // All prime numbers are 1 or 5 modulo 6.
+        // Since we start with 3, this will do:
+        // 3 -> 5 -> 7 -> 11 ... +2 -> +4 -> +2 -> +4 ...
+        i = i % 6 == 1 ? i + 4 : i + 2) {
       if (this % i == 0) return false;
     }